api_based_support_bot/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
jinja2==3.1.6
python-dotenv==1.1.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
groq==1.1.2
numpy==1.26.4
loguru==0.7.2
pytest==8.3.2
mysql-connector-python==9.1.0
requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.2
scikit-learn==1.5.0
